Develop

+++ lib/UI/views/bug_report/bug_report_view_model.dart
@@ -1,21 +1,20 @@
-import 'dart:convert';
 import 'dart:io' show Platform;
-
 import 'package:appmetrica_plugin/appmetrica_plugin.dart';
+import 'package:dio/dio.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_secure_storage/flutter_secure_storage.dart';
 import 'package:kemsu_app/UI/views/auth/auth_screen.dart';
 import 'package:kemsu_app/UI/views/bug_report/bug_report_model.dart';
 import 'package:stacked/stacked.dart';
-import 'package:http/http.dart' as http;
 import 'package:device_info_plus/device_info_plus.dart';
-
 import '../../../Configurations/config.dart';
+import '../../../domain/dio_wrapper/dio_client.dart';
 
 class BugReportViewModel extends BaseViewModel {
   BugReportViewModel(BuildContext context);
 
-  FlutterSecureStorage storage = const FlutterSecureStorage();
+  final DioClient dio = DioClient(Dio());
+  final FlutterSecureStorage storage = const FlutterSecureStorage();
   bool circle = true;
   String deviceInfoParam = "";
   List<ReportModel> reportList = [];
@@ -34,7 +33,7 @@ class BugReportViewModel extends BaseViewModel {
     AppMetrica.reportEvent('BugReport event');
   }
 
-  getDeviceInfo() async {
+  Future<void> getDeviceInfo() async {
     var deviceInfo = DeviceInfoPlugin();
     if (Platform.isAndroid) {
       AndroidDeviceInfo androidInfo = await deviceInfo.androidInfo;
@@ -52,23 +51,43 @@ class BugReportViewModel extends BaseViewModel {
     }
   }
 
-  sendAction(context) async {
+  Future<void> sendAction(context) async {
     String? accessToken = await storage.read(key: "tokenKey");
+    bool isTestUser = accessToken == 'accessToken';
+
     if (errorMsgController.text.isNotEmpty) {
       notifyListeners();
-      Map<String, String> header = {"X-Access-Token": "$accessToken"};
-
       Map<String, dynamic> body = {"message": errorMsgController.text, "deviceInfo": deviceInfoParam};
 
-      var response = await http.post(Uri.parse(Config.addReport),
-          headers: header, body: body);
+      try {
+        if (isTestUser) {
+          reportList.add(
+            ReportModel(
+              reportId: 0,
+              message: errorMsgController.text,
+              messageDate: "15.06.2024",
+              reportStatus: "Заявка не обработана",
+            ),
+          );
+          Navigator.pop(context);
+          ScaffoldMessenger.of(context).showSnackBar(const SnackBar(content: Text("Ваше обращение успешно отправлено (мок).")));
+        } else {
+          final response = await dio.post(
+            Config.addReport,
+            options: Options(headers: {'x-access-token': accessToken}),
+            data: body,
+          );
 
-      if (response.statusCode == 200 || response.statusCode == 201) {
-        await fetchReports(context);
-        Navigator.pop(context);
-        ScaffoldMessenger.of(context).showSnackBar(const SnackBar(content: Text("Ваше обращение успешно отправлено.")));
-      } else {
-        ScaffoldMessenger.of(context).showSnackBar(SnackBar(content: Text(json.decode(response.body)["message"])));
+          if (response.statusCode == 200 || response.statusCode == 201) {
+            await fetchReports(context);
+            Navigator.pop(context);
+            ScaffoldMessenger.of(context).showSnackBar(const SnackBar(content: Text("Ваше обращение успешно отправлено.")));
+          } else {
+            ScaffoldMessenger.of(context).showSnackBar(SnackBar(content: Text(response.data["message"])));
+          }
+        }
+      } catch (e) {
+        ScaffoldMessenger.of(context).showSnackBar(const SnackBar(content: Text("Произошла ошибка при отправке обращения.")));
       }
     } else {
       ScaffoldMessenger.of(context).showSnackBar(const SnackBar(content: Text("Заполните все поля обращения!")));
@@ -80,20 +99,41 @@ class BugReportViewModel extends BaseViewModel {
     AppMetrica.reportEvent('Send bugreport event');
   }
 
-  fetchReports(context) async {
+  Future<void> fetchReports(context) async {
     String? accessToken = await storage.read(key: "tokenKey");
-    Map<String, String> header = {"X-Access-Token": "$accessToken"};
-    var response =
-        await http.get(Uri.parse(Config.bugReport), headers: header);
-    if (response.statusCode == 200 || response.statusCode == 201) {
-      reportList = parseReports(json.decode(response.body)["result"]);
-      notifyListeners();
-    } else if (response.statusCode == 401) {
-      Navigator.of(context).pushAndRemoveUntil(
-          MaterialPageRoute(builder: (context) => const AuthScreen()), (Route<dynamic> route) => false);
-      await storage.delete(key: "tokenKey");
-      ScaffoldMessenger.of(context)
-          .showSnackBar(const SnackBar(content: Text("Сессия ЭИОС истекла. Пожалуйста, авторизуйтесь повторно")));
+    bool isTestUser = accessToken == 'accessToken';
+
+    try {
+      if (isTestUser) {
+        reportList = [
+          ReportModel(
+            reportId: 0,
+            message: "Тестовый отчет",
+            messageDate: "15.06.2024",
+            reportStatus: "Решено",
+          )
+        ];
+        notifyListeners();
+      } else {
+        final response = await dio.get(
+          Config.bugReport,
+          options: Options(headers: {'x-access-token': accessToken}),
+        );
+
+        if (response.statusCode == 200 || response.statusCode == 201) {
+          reportList = parseReports(response.data["result"]);
+          notifyListeners();
+        } else if (response.statusCode == 401) {
+          Navigator.of(context).pushAndRemoveUntil(
+            MaterialPageRoute(builder: (context) => const AuthScreen()),
+            (Route<dynamic> route) => false,
+          );
+          await storage.delete(key: "tokenKey");
+          ScaffoldMessenger.of(context).showSnackBar(const SnackBar(content: Text("Сессия ЭИОС истекла. Пожалуйста, авторизуйтесь повторно")));
+        }
+      }
+    } catch (e) {
+      ScaffoldMessenger.of(context).showSnackBar(const SnackBar(content: Text("Произошла ошибка при получении обращений.")));
     }
   }
